+++ udata_front/theme/gouvfr/datagouv-components/src/components/Well/Well.stories.ts
@@ -1,9 +1,12 @@
-import Well from './Well.vue';
-import editIcon from "../../../../../templates/svg/illustrations/edit.svg";
+import Well from '.';
+import type { WellProps } from './Well.vue';
+import editIcon from "../../../../templates/svg/illustrations/edit.svg";
+import type { Meta, StoryObj } from '@storybook/vue3';
 
-export default {
+const meta = {
   title: 'Ui/Well',
   component: Well,
+  tags: ['autodocs'],
   argTypes: {
     type: {
       options: [
@@ -22,13 +25,22 @@ export default {
       control: 'select',
     }
   },
-};
+  parameters: {
+    docs: {
+      description: {
+        component: "This component is a simple container with a colored background and some padding.",
+      },
+    },
+  },
+} satisfies Meta<typeof Well>;
 
-const args = {
+export default meta;
+
+const args: WellProps = {
   type: "primary",
 };
 
-export const SimpleWell = {
+export const SimpleWell: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { Well },
     setup() {
@@ -43,16 +55,16 @@ export const SimpleWell = {
   args,
 };
 
-const argsRegular = {
+const argsRegular: WellProps = {
   color: "blue-cumulus",
   weight: "regular"
 };
 
-export const WellWithImg = {
+export const WellWithImg: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { Well },
     setup() {
-      return { args: argsRegular, editIcon };
+      return { args, editIcon };
     },
     template: ` <Well v-bind="args">
                     <div class="fr-grid-row">
@@ -66,5 +78,5 @@ export const WellWithImg = {
                     </div>
                   </Well>`,
   }),
-  args,
+  args: argsRegular,
 }