+++ udata_front/theme/gouvfr/datagouv-components/src/components/index.ts
@@ -2,5 +2,6 @@ import { CopyButton } from "./CopyButton";
 import OrganizationNameWithCertificate from "./Organization";
 import Pagination from "./Pagination";
 import { ResourceAccordion, Loader as ResourceAccordionLoader } from "./ResourceAccordion";
+import Well from "./Well";
 
-export { CopyButton, OrganizationNameWithCertificate, Pagination, ResourceAccordion, ResourceAccordionLoader };
+export { CopyButton, OrganizationNameWithCertificate, Pagination, ResourceAccordion, ResourceAccordionLoader, Well };